cleanup + static changelog generation
